+++ CHANGES.md
@@ -6,6 +6,7 @@
  * Feature: Access to all AIOKafka configuration options
  * Feature: Use backend Queue for Kafka
  * Feature: Add support for storing book snapshots in Redis as key-value
+ * Update: Switch from unmaintained aioredis to redis-py
 
 ### 2.3.1 (2022-10-31)
  * Bugfix: timestamp not reset correctly on reconnect